+++ src/components/forms/DatePickerField/styles.ts
@@ -10,3 +10,15 @@ export const datePickerContainerSmall = css`
   ${datePickerContainer};
   width: unset;
 `
+
+export const calendarButton = css`
+  position: absolute;
+  right: 5px;
+  top: 9px;
+  z-index: 2;
+`
+
+export const calendarButtonSmall = css`
+  right: 0;
+  top: 0;
+`