+++ src/components/layout/Header/Header.tsx
@@ -3,7 +3,9 @@ import React, { Children, isValidElement, cloneElement, ReactElement } from 'rea
 import { jsx } from '@emotion/core'
 
 import { Color, R400_COLOR } from '../../../colors/index'
+
 import { FerpectionLogo } from '../../iconography/FerpectionLogo'
+import { HeaderLink } from '../../buttons/HeaderLink'
 
 import { defaultHeaderStyles, buttonListStyle, lightHeaderStyle, logoStyles } from './styles'
 
@@ -18,13 +20,18 @@ export const Header: React.FC<HeaderProps> = props => {
 
   const buttonList = Children.toArray(children)
     .filter(child => isValidElement(child))
-    .map((child: ReactElement) =>
-      cloneElement(child, {
+    .map((child: ReactElement) => {
+      const opts: any = {
         isFilled: false,
         color: R400_COLOR,
-        headerLigthVersion: isLightVersion,
-      }),
-    )
+      }
+
+      if (child.type === HeaderLink) {
+        opts.headerLigthVersion = isLightVersion
+      }
+
+      return cloneElement(child, opts)
+    })
     .map((child, i) => <li key={i}>{child}</li>)
 
   return (