+++ CHANGELOG.md
@@ -5,8 +5,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) since [v0.1.0].
 
 ## [Unreleased]
+### Added
+- `RadioGroup` now supports errors & validation:
+  A `RadioGroup` have an visual error state and can be marked as required.
+  It is also possible to add custom validators to influence how the component
+  triggers the error state.
+- `CheckboxGroup` now supports errors & validation:
+  A `CheckboxGroup` have an visual error state and can be marked as required.
+  It is also possible to add custom validators to influence how the component triggers
+  the error state.
 
-## [v0.6.0] - 2020-10-16
+### Changed
+- Improve typing experience in the `DatePickerField` component. Users can show and hide
+  the calendar using a button at the right. Users can type a text, when it becomes a
+  valid date it is used in the calendar.
+
+### Fixed
+- Remove an internal error displayed in the console when a `Button` was used directly in
+  the `Header` component.
+
+## [v0.6.0] - 2020-10-16s
 ### Added
 - New `validators` attribute for all the form components to display custom errors.
 - The `DatePicker` component allows for direct month and year selection.